Added PatternBase and MatchingResultBase.
