+++ Sources/PT.PM.Patterns/Pattern.cs
@@ -1,54 +1,19 @@
-﻿using PT.PM.Common;
-using PT.PM.Common.Nodes;
-using PT.PM.Patterns.Nodes;
-using Newtonsoft.Json;
-using System;
+﻿using PT.PM.Patterns.Nodes;
 
 namespace PT.PM.Patterns
 {
-    public class Pattern
+    public class Pattern : PatternBase
     {
-        private LanguageFlags languages = LanguageExt.AllPatternLanguages;
-
-        public string Key { get; set; }
-
-        public LanguageFlags Languages
-        {
-            get
-            {
-                return languages;
-            }
-            set
-            {
-                if (value.Is(LanguageFlags.Aspx))
-                {
-                    throw new ArgumentException($"Unable to create pattern for Aspx");
-                }
-                languages = value;
-            }
-        }
-
         public PatternNode Data { get; set; }
 
         public Pattern(PatternDto patternDto, PatternNode data)
+            : base(patternDto.Key, patternDto.Description, patternDto.Languages)
         {
-            Key = patternDto.Key;
-            DebugInfo = patternDto.Description;
-            Languages = patternDto.Languages;
-
             Data = data;
         }
 
-        [JsonIgnore]
-        public string DebugInfo { get; set; }
-
         public Pattern()
         {
-        }
-
-        public override string ToString()
-        {
-            return DebugInfo;
         }
     }
 }
